# docs/swagger.yaml
definitions:
  handler.accountContractResponse:
    properties:
      description:
        type: string
      fee:
        type: integer
      id:
        type: integer
      imageUrl:
        type: string
      isMain:
        type: boolean
      name:
        type: string
      type:
        type: string
    type: object
...
  handler.completeAccountRequest:
    properties:
      id:
        type: integer
      status:
        type: string
    type: object
...
  handler.errorResponse:
    properties:
      error:
        type: string
    type: object
  handler.getAccountListRequest:
    properties:
      from:
        type: string
      status:
        type: string
    type: object
  handler.getAccountListResponse:
    properties:
      accounts:
        items:
          $ref: '#/definitions/handler.getAccountsListAccount'
        type: array
    type: object
  handler.getAccountResponse:
    properties:
      contracts:
        items:
          $ref: '#/definitions/handler.accountContractResponse'
        type: array
      createdAt:
        type: string
      creator:
        type: integer
      finishedAt:
        type: string
      id:
        type: integer
      moderator:
        type: integer
      number:
        type: string
      requestedAt:
        type: string
      status:
        type: string
      totalFee:
        type: integer
    type: object
  handler.getAccountsListAccount:
    properties:
      createdAt:
        type: string
      creator:
        type: integer
      finishedAt:
        type: string
      id:
        type: integer
      moderator:
        type: integer
      number:
        type: string
      requestedAt:
        type: string
      status:
        type: string
      totalFee:
        type: integer
    type: object
...
  handler.updateAccountRequest:
    properties:
      id:
        type: integer
      number:
        type: string
    type: object
...
host: localhost:8000
info:
  contact: {}
  title: BMSTU Web Backend
  version: "1.0"
paths:
  /account/{accountId}/complete:
    post:
      consumes:
      - application/json
      description: Завершает заявку на счёт, обновляя её статус по ID заявки.
      parameters:
      - description: ID заявки на счёт
        in: path
        name: accountId
        required: true
        type: integer
      - description: Данные для завершения заявки
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.completeAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Заявка на счёт успешно завершена
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Заявка на счёт не найдена
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - Bearer: []
      summary: Завершение заявки на счёт
      tags:
      - accounts
  /account/{accountId}/submit:
    post:
      consumes:
      - application/json
      description: Отправляет заявку на счёт по её ID для дальнейшей обработки.
      parameters:
      - description: ID заявки на счёт
        in: path
        name: accountId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Заявка на счёт успешно отправлена
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Заявка на счёт не найдена
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - Bearer: []
      summary: Отправка заявки на счёт
      tags:
      - accounts
  /accounts:
    get:
      consumes:
      - application/json
      description: Возвращает список всех заявок на счёт с возможностью фильтрации
        по статусу и дате.
      parameters:
      - description: Фильтры для получения заявок на счёт
        in: body
        name: request
        schema:
          $ref: '#/definitions/handler.getAccountListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Список заявок на счёт
          schema:
            $ref: '#/definitions/handler.getAccountListResponse'
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - Bearer: []
      summary: Получение списка заявок на счёт
      tags:
      - accounts
  /accounts/{accountId}:
    delete:
      consumes:
      - application/json
      description: Удаляет заявку на счёт по её ID.
      parameters:
      - description: ID заявки на счёт
        in: path
        name: accountId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Заявка на счёт успешно удалена
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Заявка на счёт не найдена
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - Bearer: []
      summary: Удаление заявки на счёт
      tags:
      - accounts
    get:
      consumes:
      - application/json
      description: Возвращает подробную информацию о конкретной заявке на счёт по
        её ID, включая связанные договоры.
      parameters:
      - description: ID заявки на счёт
        in: path
        name: accountId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Детали заявки на счёт
          schema:
            $ref: '#/definitions/handler.getAccountResponse'
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Заявка на счёт не найдена
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - Bearer: []
      summary: Получение информации о заявке на счёт
      tags:
      - accounts
    put:
      consumes:
      - application/json
      description: Обновляет номер заявки на счёт по её ID.
      parameters:
      - description: ID заявки на счёт
        in: path
        name: accountId
        required: true
        type: integer
      - description: Данные для обновления заявки
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.updateAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Заявка на счёт успешно обновлена
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Заявка на счёт не найдена
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - Bearer: []
      summary: Обновление информации о заявке на счёт
      tags:
      - accounts
  /accounts/{accountId}/contract/{contractId}:
    delete:
      consumes:
      - application/json
      description: Удаляет указанный договор из заявки на счёт по их ID.
      parameters:
      - description: ID заявки на счёт
        in: path
        name: accountId
        required: true
        type: integer
      - description: ID договора
        in: path
        name: contractId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Договор успешно удалён из заявки на счёт
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Заявка на счёт или договор не найдены
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - Bearer: []
      summary: Удаление договора из заявки на счёт
      tags:
      - account-contracts
  /accounts/{accountId}/contract/{contractId}/main:
    post:
      consumes:
      - application/json
      description: Устанавливает указанный договор в качестве основного для заявки
        на счёт по их ID.
      parameters:
      - description: ID заявки на счёт
        in: path
        name: accountId
        required: true
        type: integer
      - description: ID договора
        in: path
        name: contractId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Договор успешно установлен как основной в заявке на счёт
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Заявка на счёт или договор не найдены
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - Bearer: []
      summary: Установить договор основным в заявке на счёт
      tags:
      - account-contracts
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"